% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_url.R
\name{cache_url}
\alias{cache_url}
\title{Cache URL content}
\usage{
cache_url(url, debug = FALSE)
}
\arguments{
\item{url}{The URL to fetch the content from.}

\item{debug}{When \code{TRUE} will emit useful messages telling you about
cache hits and misses. This can be helpful to understand whether or
not caching is actually doing anything for your use case.}
}
\value{
The content of the URL, either fetched as a string. If the provided path is not a url, it is returned as is.
}
\description{
This function fetches the content of a given URL and caches it for future use.
If the provided path is not an URL, it returns the path so it can be used directly by reading functions.
}
