% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current.R
\name{get_current}
\alias{get_current}
\title{Get the version of the dataset as currently present locally. Called current
in tests.}
\usage{
get_current()
}
\value{
A data.frame with DwC mapping of the dataset, as present on the repo.
}
\description{
This version might be different from the reference dataset \code{\link[=get_reference]{get_reference()}}
if new data was fetched and mapped but not yet published to Github. For
example during an automatic publishing run, or while a new verion of the
dataset is still in review.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_current()
\dontshow{\}) # examplesIf}
}
