% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_modified_lines.R
\name{get_modified_lines}
\alias{get_modified_lines}
\title{Get the lines that changed between two DwC occurrence files}
\usage{
get_modified_lines(
  current_path = "data/processed/occurrence.csv",
  reference_path =
    paste0("https://raw.githubusercontent.com/riparias/rato-occurrences/main/",
    "data/processed/occurrence.csv"),
  drop_new_records = TRUE,
  as_df = TRUE
)
}
\arguments{
\item{current_path}{The current file path to compare against. By default, this is the
file in the \code{data/processed} directory.}

\item{reference_path}{The reference file path to compare against. By default, this is
the file in the GitHub repository.}

\item{drop_new_records}{Should new records (based on \code{occurrenceID}) be
excluded from the result?}

\item{as_df}{By default, return the modified lines as a data frame. If FALSE,
return as a character vector of lines.}
}
\value{
A data frame with the modified lines or a character vector of lines.
}
\description{
Optionally you can exclude new records (based on \code{occurrenceID}) from the
result. If \code{drop_new_records} is set to \code{FALSE}, this function should work
for any two text files. It's based on \code{spookyhash} hashes compared line by
line.
}
\examples{
get_modified_lines()
}
