% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_objects.R
\name{get_objects}
\alias{get_objects}
\title{Get objects from the RATO ArcGIS Enterprise environment via the API}
\usage{
get_objects(object_ids, token = get_token())
}
\arguments{
\item{object_ids}{Vector, will be converted to character. Of object_ids}

\item{token}{Access token, by default generated by \code{get_token()}}
}
\value{
data.frame
}
\description{
Get objects from the RATO ArcGIS Enterprise environment via the API
}
