% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reference.R
\name{get_reference}
\alias{get_reference}
\title{Get the version of the dataset as published on Github, called reference in
tests}
\usage{
get_reference()
}
\value{
A data.frame with DwC mapping of the dataset, as currently published
on Github.
}
\description{
This version might be different from the one on GBIF, as the IPT only fetches
the data from GITHUB periodically (once a week) and not necessarily in phase
with the updates on Github.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_reference()
\dontshow{\}) # examplesIf}
}
