% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_token.R
\name{get_token}
\alias{get_token}
\title{Request an access token from the GIS API services}
\usage{
get_token(
  username = "RATO_INBO",
  password = Sys.getenv("ratopwd"),
  expires = 5
)
}
\arguments{
\item{username}{ArcGIS Enterprise username}

\item{password}{ArcGIS Enterprise password}

\item{expires}{In minutes, how long should the token remain valid?}
}
\value{
Character. An access token for future API calls.
}
\description{
Request an access token from the GIS API services
}
