% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_modified_lines.R
\name{line_which}
\alias{line_which}
\title{Find the line numbers in a vector that match fixed patterns}
\usage{
line_which(x, fixed_patterns)
}
\arguments{
\item{x}{A character vector to search for patterns}

\item{fixed_patterns}{A character vector of fixed patterns to match}
}
\value{
A numeric vector of line numbers where the patterns are found
}
\description{
This function takes a vector \code{x} and a vector of \code{fixed_patterns},
and returns the line numbers in \code{x} where the patterns are found.
}
\examples{
x <- c("a", "b", "c", "d", "e")
fixed_patterns <- c("b", "d")
line_which(x, fixed_patterns)
}
