% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_object_ids.R
\name{list_object_ids}
\alias{list_object_ids}
\title{List all object ids in the RATO ArcGIS Enterprise MapServer}
\usage{
list_object_ids(
  token = get_token(),
  domain = "https://gis.oost-vlaanderen.be/server/rest/services/",
  table_path = c("RATO2", "RATO2_Dossiers_Publiek", "MapServer"),
  query = c("0", "query?where=1\%3D1")
)
}
\arguments{
\item{token}{Access token, by default generated by \code{get_token()}}

\item{domain}{URL to the ArcGIS Server GIS services}

\item{table_path}{Vector of strings, the path to the table in the MapServer}

\item{query}{Vector of strings, the query to the table in the MapServer}
}
\value{
Integer vector of (all) object ids.
}
\description{
List all object ids in the RATO ArcGIS Enterprise MapServer
}
