% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_dwc_order.R
\name{set_dwc_order}
\alias{set_dwc_order}
\title{Set the order of the columns to a specific order.}
\usage{
set_dwc_order(x)
}
\arguments{
\item{x}{Input tibble as produced by the dwc mapping}
}
\value{
A tibble with the columns in the default order of the dwc mappings on
the IPT, except identiable columns are placed at the front.
}
\description{
This order is relevant because it is easier to read in a text editor, as well
as increasing the chances that git will correctly detect a change, addition
or removal based on the identifiable columns being at the front.
}
\details{
A notable difference from default DWC-A field order is that identifing fields
are placed at the front to increase the chances of GIT correctly detecting
changes as additions, changes or deletions. Columns that are less likely to
change are placed at the back.
}
