% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species}
\alias{species}
\title{A vector of species names known to be included in the RATO dataset}
\format{
\code{species} A vector with over 30 species names.
}
\usage{
species
}
\description{
This vector contains a manually created reference list of species names known
to be included in this dataset. Checks are run against this list to detect
when new species are added.
}
\keyword{datasets}
