% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_encrypted_csv.R
\name{write_encrypted_csv}
\alias{write_encrypted_csv}
\title{Write an encrypted csv}
\usage{
write_encrypted_csv(x, outfile, key = Sys.getenv("encryption_key"), ...)
}
\arguments{
\item{x}{A data frame or tibble to write to disk in an encrypted binary file.}

\item{outfile}{Non-existant file where the decrypted output is to be written}

\item{key}{For symmetric decryption, 'pkey' should be NULL (default) and
'key' can be either a string (Default is 'pass') or a raw object. For
asymmetric decryption, both 'key' (private key of the decrypter) and 'pkey'
(public key of the encrypter) should be raw objects.}

\item{...}{passed on to \code{\link[readr:read_delim]{readr::read_csv()}}}
}
\value{
A binary file, the function returns TRUE invisibly on success.
}
\description{
Write an encrypted csv
}
