% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_object_ids.R
\name{list_object_ids}
\alias{list_object_ids}
\title{List all object ids in the RATO ArcGIS Enterprise MapServer}
\usage{
list_object_ids(token = get_token())
}
\arguments{
\item{token}{Access token, by default generated by \code{get_token()}}
}
\value{
Integer vector of (all) object ids.
}
\description{
List all object ids in the RATO ArcGIS Enterprise MapServer
}
